🚚 Relocate study info steps into its own file
