+++ src/forms/StudyInfoForm/Steps/LogisticsStep.js
@@ -6,236 +6,16 @@ import {
   Button,
   Confirm,
   List,
-  Dropdown,
   Header,
   Image,
   Table,
 } from 'semantic-ui-react';
-import {workflowOptions} from '../common/enums';
+import {workflowOptions} from '../../../common/enums';
+import {prevNextStep} from '../../../common/notificationUtils';
 import FormField from './FormField';
 import Markdown from 'react-markdown';
 
-const prevNextStep = (stepName, newStudy, history) => {
-  if (newStudy) {
-    history.push('/study/new-study/' + stepName);
-  } else {
-    history.push(
-      '/study/' +
-        history.location.pathname.split('/')[2] +
-        '/basic-info/' +
-        stepName,
-    );
-  }
-};
-
-export const InfoStep = ({
-  formikProps,
-  setActiveStep,
-  setFocused,
-  focused,
-  setSelection,
-  workflowType,
-  newStudy,
-  history,
-  editing,
-  isAdmin,
-}) => {
-  const {values, errors, touched, handleChange, handleBlur} = formikProps;
-  return (
-    <>
-      <Header
-        as="h4"
-        className="text-wrap-75"
-        content="Please provide the study's full name and a shortened version that may be used for display purposes."
-      />
-      <FormField
-        isAdmin={isAdmin}
-        newStudy={newStudy}
-        required
-        id="name"
-        name="Study Name"
-        description="Full name of the study, often the full title of the X01 grant application."
-        focused={focused === 'name'}
-        value={values.name}
-        touched={touched.name}
-        errors={errors.name}
-        handleChange={handleChange}
-        handleBlur={handleBlur}
-        handleFocus={id => setFocused(id)}
-        descriptionDisplay="tooltip"
-        readOnly={!editing && !newStudy}
-      />
-      <FormField
-        isAdmin={isAdmin}
-        newStudy={newStudy}
-        id="shortName"
-        name="Study Short Name"
-        description="The name that will appear under portal facets."
-        type="text"
-        focused={focused === 'shortName'}
-        value={values.shortName}
-        touched={touched.shortName}
-        errors={errors.shortName}
-        handleChange={handleChange}
-        handleBlur={handleBlur}
-        handleFocus={id => setFocused(id)}
-        readOnly={!editing && !newStudy}
-      />
-      {newStudy ? (
-        <FormField
-          isAdmin={isAdmin}
-          newStudy={newStudy}
-          id="workflowType"
-          name="Cavatica Projects"
-          description="Workflow projects to be instantiated for the new study."
-          focused={focused === 'workflowType'}
-          value={values.workflowType}
-          touched={touched.workflowType}
-          errors={errors.workflowType}
-          handleChange={handleChange}
-          handleBlur={handleBlur}
-          handleFocus={id => setFocused(id)}
-        >
-          <Dropdown
-            id="workflowType"
-            name="workflowType"
-            placeholder="Workflow Type"
-            fluid
-            selection
-            clearable
-            multiple
-            options={workflowOptions}
-            onChange={(e, {value}) => {
-              setSelection(value);
-            }}
-            value={workflowType}
-          />
-        </FormField>
-      ) : (
-        <FormField
-          isAdmin={isAdmin}
-          newStudy={newStudy}
-          id="bucket"
-          name="S3 Bucket"
-          description="The s3 bucket where data for this study resides."
-          type="text"
-          focused={focused === 'bucket'}
-          value={values.bucket}
-          touched={touched.bucket}
-          errors={errors.bucket}
-          handleChange={handleChange}
-          handleBlur={handleBlur}
-          handleFocus={id => setFocused(id)}
-          readOnly={!editing && !newStudy}
-        />
-      )}
-      <Button
-        primary
-        floated="right"
-        type="button"
-        onClick={() => prevNextStep('external', newStudy, history)}
-        labelPosition="right"
-        icon="right arrow"
-        content="NEXT"
-        disabled={values.name.length === 0}
-      />
-    </>
-  );
-};
-
-export const ExternalStep = ({
-  formikProps,
-  setActiveStep,
-  setFocused,
-  focused,
-  editing,
-  newStudy,
-  history,
-  isAdmin,
-}) => {
-  const {values, errors, touched, handleChange, handleBlur} = formikProps;
-  return (
-    <>
-      <Header
-        as="h4"
-        className="text-wrap-75"
-        content="If the study is a dbGaP project, additional information is needed to ensure access is granted correctly. For non-dbGaP studies, an external identifier which this study may otherwise be known by is required."
-      />
-      <FormField
-        isAdmin={isAdmin}
-        newStudy={newStudy}
-        required
-        id="externalId"
-        name="External ID"
-        description="Identifier used by external systems, often the PHS ID if the study is registered with dbGaP."
-        type="text"
-        focused={focused === 'externalId'}
-        placeholder="Example: phs000178"
-        value={values.externalId}
-        touched={touched.externalId}
-        errors={errors.externalId}
-        handleChange={handleChange}
-        handleBlur={handleBlur}
-        handleFocus={id => setFocused(id)}
-        readOnly={!editing && !newStudy}
-      />
-      <FormField
-        isAdmin={isAdmin}
-        newStudy={newStudy}
-        id="version"
-        name="dbGaP Version"
-        placeholder="Example: v1.p1"
-        description="Study version, often the provided by the data access authority."
-        type="text"
-        focused={focused === 'version'}
-        value={values.version}
-        touched={touched.version}
-        errors={errors.version}
-        handleChange={handleChange}
-        handleBlur={handleBlur}
-        handleFocus={id => setFocused(id)}
-        readOnly={!editing && !newStudy}
-      />
-      <FormField
-        isAdmin={isAdmin}
-        newStudy={newStudy}
-        id="attribution"
-        name="Attribution"
-        description="The URL providing more information about the study."
-        type="text"
-        focused={focused === 'attribution'}
-        value={values.attribution}
-        touched={touched.attribution}
-        errors={errors.attribution}
-        handleChange={handleChange}
-        handleBlur={handleBlur}
-        handleFocus={id => setFocused(id)}
-        readOnly={!editing && !newStudy}
-      />
-      <Button
-        primary
-        floated="left"
-        type="button"
-        onClick={() => prevNextStep('info', newStudy, history)}
-        labelPosition="left"
-        icon="left arrow"
-        content="PREVIOUS"
-      />
-      <Button
-        primary
-        floated="right"
-        type="button"
-        onClick={() => prevNextStep('logistics', newStudy, history)}
-        labelPosition="right"
-        icon="right arrow"
-        content="NEXT"
-        disabled={values.externalId.length === 0}
-      />
-    </>
-  );
-};
-
-export const LogisticsStep = ({
+const LogisticsStep = ({
   newStudy,
   formikProps,
   setActiveStep,
@@ -459,3 +239,5 @@ export const LogisticsStep = ({
     </>
   );
 };
+
+export default LogisticsStep;
