+++ src/forms/StudyInfoForm/Steps/index.js
@@ -1,0 +1,3 @@
+export {default as InfoStep} from './InfoStep';
+export {default as LogisticsStep} from './LogisticsStep';
+export {default as ExternalStep} from './ExternalStep';
