+++ src/forms/StudyInfoForm/index.js
@@ -1,0 +1,1 @@
+export {default as NewStudyForm} from './NewStudyForm';
